Do the `<Leader>t` thing
The Makefile has been made a bit more 'generic' so that, for instance,
`make 01-1` will compile `src/01-1.c` to `01-1.o`.
A vim command is supplied that'll map our beloved `<Leader>t` to a
series of commands that will, based on the file in the current buffer,
compile the current file and execute its executable.